Add support for search progress output
Add a new --show-progress option, which controls the output of search
progress during a search.  Each file being searched may be displayed as
a progress update; a new thread takes care of displaying progress
updates up to 20 times a second.  Progress is only shown in a terminal.
 Note that due to the additional tracking and updates required,
enabling --show-progress can make large searches a few percent slower,
but it may be perceived to be faster due to the progress output.